close #259

+++ db/migrate/20190611091036_remove_gcm_token_from_users.rb
@@ -1,7 +1,5 @@
 class RemoveGcmTokenFromUsers < ActiveRecord::Migration[5.2]
   def change
-    if column_exists?(:users, :gcm_token)
-      remove_column :users, :gcm_token
-    end
+    remove_column :users, :gcm_token if column_exists?(:users, :gcm_token)
   end
 end
